requirements.txt
Flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
openai==1.3.7
numpy==1.24.3
scipy==1.11.4
gunicorn==21.2.0

